switched appolo provider to App.js, tried to get graphql data to frontend, not working yet
